Move data update coordinators to the coordinator module
