+++ homeassistant/components/nextdns/button.py
@@ -9,8 +9,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import NextDnsStatusUpdateCoordinator
 from .const import ATTR_STATUS, DOMAIN
+from .coordinator import NextDnsStatusUpdateCoordinator
 
 PARALLEL_UPDATES = 1
